Refactored flag_utils.py away
